+++ src/main/java/sic/vista/swing/UsuariosGUI.java
@@ -1,5 +1,6 @@
 package sic.vista.swing;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.ResourceBundle;
@@ -78,7 +79,7 @@ public class UsuariosGUI extends JInternalFrame {
         tbl_Resultado.getTableHeader().setResizingAllowed(true);
 
         //render para los tipos de datos
-        tbl_Resultado.setDefaultRenderer(Double.class, new RenderTabla());
+        tbl_Resultado.setDefaultRenderer(BigDecimal.class, new RenderTabla());
 
         //tamanios de columnas
         tbl_Resultado.getColumnModel().getColumn(0).setPreferredWidth(400);
